modified facilities manager for director role

+++ AviatoCore.Web/Controllers/FacilitiesController.cs
@@ -3,6 +3,7 @@ using AviatoCore.Domain.Entities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 [Route("api/[controller]")]
@@ -18,16 +19,16 @@ public class FacilitiesController : ControllerBase
 
     // GET: api/Facilities 
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<Facility>>> GetFacilities()
+    public async Task<ActionResult<IEnumerable<Facility>>> GetFacilities([FromQuery] int airportId)
     {
-        return Ok(await _facilityService.GetAllFacilitiesAsync());
+        return Ok(await _facilityService.GetFacilitiesByAirportIdAsync(airportId));
     }
 
     // GET: api/Facilities/5
     [HttpGet("{id}")]
-    public async Task<ActionResult<Facility>> GetFacility(int id)
+    public async Task<ActionResult<Facility>> GetFacility(int id, [FromQuery] int airportId)
     {
-        var facility = await _facilityService.GetFacilityAsync(id);
+        var facility = await _facilityService.GetFacilityAsync(id, airportId);
 
         if (facility == null)
         {
@@ -38,11 +39,11 @@ public class FacilitiesController : ControllerBase
     }
 
     // PUT: api/Facilities/5
-    [Authorize(Roles = "Admin")]
+    [Authorize(Roles = "Director")]
     [HttpPut("{id}")]
-    public async Task<IActionResult> PutFacility(int id, Facility facility)
+    public async Task<IActionResult> PutFacility(int id, [FromQuery] int airportId, Facility facility)
     {
-        if (id != facility.Id)
+        if (id != facility.Id || airportId != facility.AirportId)
         {
             return BadRequest();
         }
@@ -53,20 +54,32 @@ public class FacilitiesController : ControllerBase
     }
 
     // POST: api/Facilities
-    [Authorize(Roles = "Admin")]
+    [Authorize(Roles = "Director")]
     [HttpPost]
-    public async Task<ActionResult<Facility>> PostFacility(Facility facility)
+    public async Task<ActionResult<Facility>> PostFacility([FromQuery] int airportId, Facility facility)
     {
+        if (airportId != facility.AirportId)
+        {
+            return BadRequest();
+        }
+
         await _facilityService.AddFacilityAsync(facility);
 
         return CreatedAtAction("GetFacility", new { id = facility.Id }, facility);
     }
 
     // DELETE: api/Facilities/5
-    [Authorize(Roles = "Admin")]
+    [Authorize(Roles = "Director")]
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteFacility(int id)
+    public async Task<IActionResult> DeleteFacility(int id, [FromQuery] int airportId)
     {
+        var facility = await _facilityService.GetFacilityAsync(id, airportId);
+
+        if (facility == null || facility.AirportId != airportId)
+        {
+            return Unauthorized();
+        }
+
         await _facilityService.DeleteFacilityAsync(id);
 
         return NoContent();
